Remove psutil runtime dependency
